test for default reseller assignment
