+++ .gitignore
@@ -5,3 +5,4 @@
 /node_modules
 node_modules/
 .idea/
+.env
