+++ test/.env
@@ -6,9 +6,11 @@ export SAUCE_USERNAME=
 export SAUCE_ACCESS_KEY=
 
 
-DB_HOST=mysql2018.pagevamp.com
-DB_USERNAME=kusum42384
-DB_PASSWORD=P72D#!NKknE^PILggUw5
+DB_HOST=mysql
+DB_USERNAME=root
+DB_PASSWORD=root
 DB_DATABASE=pagevamp
 DB_PORT=3306
+
+URL=https://www.pagevamp.com/
