+++ test/nightwatch.json
@@ -2,7 +2,7 @@
  "globals_path": "globals.js",
  "output_folder": "reports",
  "src_folders": ["specs"],
- "skip_testcases_on_fail" :false,
+ "skip_testcases_on_fail" :true,
  "test_settings": {
    "default": {
      "screenshots": {
