+++ test/package.json
@@ -6,9 +6,12 @@
   "dependencies": {
     "axios": "^0.18.0",
     "babel-core": "^6.26.3",
+    "babel-runtime": "^6.26.0",
+    "core-js": "^2.5.7",
     "env2": "^2.1.1",
     "ffmpeg": "0.0.4",
     "mysql": "^2.15.0",
+    "namshi-node-mysql": "^3.0.0",
     "nightwatch-video-recorder": "1.1.0",
     "node-tesseract": "^0.2.7",
     "promise": "^8.0.2"
@@ -16,6 +19,7 @@
   "devDependencies": {
     "babel-core": "^6.26.3",
     "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-preset-es2015": "^6.24.1"
+    "babel-preset-es2015": "^6.24.1",
+    "babel-register": "^6.26.0"
   }
 }
